[DF-312] Fixed warnings

+++ Sources/Internal/Render/Texture.h
@@ -306,7 +306,7 @@ private:
     static void SetPixelDescription(PixelFormat index, const String &name, int32 size, GLenum type, GLenum format, GLenum internalFormat);
     
 #if defined(__DAVAENGINE_OPENGL__)
-    static GLint HWglConvertWrapMode(TextureWrap wrap);
+    static GLfloat HWglConvertWrapMode(TextureWrap wrap);
 #endif //#if defined(__DAVAENGINE_OPENGL__)
     
     void ReloadAsPng();
